fix(config): merge globalExtends before default config
